requirements.txt
langchain
langchain_community
langchain_openai
python-dotenv
langgraph
langchainhub
langchain_experimental
pytest
pandas
streamlit
plotly
streamlit-plotly-events